vector: move Pick and ReversePick to pick.go
They live in view.go for historical reasons but don't really belong
there.
